+++ sdk/storage/azure_storage_blob/src/clients/blob_container_client.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-pub use crate::generated::clients::BlobContainerClient;
+pub use crate::generated::clients::{BlobContainerClient, BlobContainerClientOptions};
 
 use crate::{
     logging::apply_storage_logging_defaults, models::StorageErrorCode,
@@ -10,32 +10,13 @@ use crate::{
 use azure_core::{
     credentials::TokenCredential,
     error::ErrorKind,
-    fmt::SafeDebug,
     http::{
         policies::{auth::BearerTokenAuthorizationPolicy, Policy},
-        ClientOptions, Pipeline, StatusCode, Url,
+        Pipeline, StatusCode, Url,
     },
     tracing, Result,
 };
 use std::sync::Arc;
-
-/// Options used when creating a [`BlobContainerClient`].
-#[derive(Clone, SafeDebug)]
-pub struct BlobContainerClientOptions {
-    /// Allows customization of the client.
-    pub client_options: ClientOptions,
-    /// Specifies the version of the operation to use for this request.
-    pub version: String,
-}
-
-impl Default for BlobContainerClientOptions {
-    fn default() -> Self {
-        Self {
-            client_options: ClientOptions::default(),
-            version: String::from("2026-04-06"),
-        }
-    }
-}
 
 impl BlobContainerClient {
     /// Creates a new BlobContainerClient, using Entra ID authentication.